+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 ## CHANGE LOG ##
 
-* v1.0.1 (2020-12-02)
+* @dev
+   * Added support for desktop notifications
    * Added `Configuration` state indicating lack of widget required configuration.
    * Improved API response handling.
    * Scaled state icons down to reduce plasmoid size